Change the "full services" approach to a more POSIX one.
We now have only one Windows service in charge of starting/stopping
the relevant NUT processes. It's a Windows replacement for the init
script.

This service is also a proxy for processes to write to Windows event
log. So only one service has to be registered to be able to log event.

Fossil-ID: SVN r2704